[#23601] YSQL: Fix TestPreparedStatements tests with connection manager enabled
Summary:
Commit 4f94526eaa1f7918d1cb87d18522990f33d9c605 fixed PgSequence tests with connection manager enabled but caused `TestPreparedStatements` tests to fail.

Reason for that was `TestPreparedStatements` was creating table with `SERIAL` column which is a sequence object only and inserting a row in that table will make session sticky. This test is creating 100 logical clients with `ysql_conn_mgr_max_conns_per_db` flag value set to 100/5 = 20. Due to which only first 20 prepared statements are getting executed but the next request (21st) is not getting any backend for execution causing test to timeout.

To fix this test, serial column is removed from the table.
Jira: DB-12517

Test Plan:
Jenkins: java only

  ./yb_build.sh release --java-test org.yb.ysqlconnmgr.TestPreparedStatements

Reviewers: asrinivasan, rbarigidad, mkumar

Reviewed By: asrinivasan

Differential Revision: https://phorge.dev.yugabyte.com/D37558