[#23668] YSQL: Handle OBJECT_NOT_EXISTS errors with Connection Manager enabled
Summary:
This diff fixes tests in which we may run into OBJECT_NOT_FOUND errors as a result of attempting to execute a SELECT/UPDATE/DELETE soon after a TRUNCATE operation, when Connection Manager is enabled. If the physical connection changes between these operations, it could lead to a state of inconsistency due to the delay in cache invalidation. For the time being, we explicitly allow cache invalidation to occur in the tests, and this issue will be taken up via code changes as a part of [[ https://yugabyte.atlassian.net/browse/DB-12518 | DB-12518 ]]

Jira: DB-12579

Test Plan:
Jenkins: enable connection manager, all tests

Ensure that all modified tests are passing, and that no unexpected failures have crept in

Reviewers: asrinivasan, skumar, mkumar, stiwary, devansh.saxena

Reviewed By: asrinivasan

Differential Revision: https://phorge.dev.yugabyte.com/D37571