[PLAT-12226] Add connection pooling status to universe health check
Summary:
This diff adds health checks for connection pooling enabled universes. Following is the logic for different combinations of universe configurations:

```
1. For CP + no YSQL auth case:
- Check for odyssey uptime
- Check for YSQL port connectivity
- Check for internal YSQL port connectivity

2. For CP + YSQL auth case:
- Check for odyssey uptime
- Check for internal YSQL port connectivity
(We don't check for external port connectivity since we connect to the socket instead when YSQL auth is enabled)

3. For no CP case:
- Check for YSQL port connectivity
(Same as before this diff)
```

Test Plan:
Manually tested that the health checks are working properly for above 3 configurations of universes.
Run UTs.
Run itests.

Reviewers: vbansal

Reviewed By: vbansal

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D37573