[pg15] fix: Set index unchanged hint to false for YB relations
Summary:
pg15 merge commit be8504df264 introduced optimizations to skip secondary index updates when index columns remain unchanged.
This commit introduced a bug that assigns an incorrect value to a variable `indexUnchanged`.
This variable `indexUnchanged` is a part of a mechanism introduced by PG to pass on a hint to the storage layer that an index tuple may be unchanged.
In this revision, this variable is changed to always evaluate to false for Yugabyte relation.

Test Plan: Run Jenkins

Reviewers: jason

Reviewed By: jason

Subscribers: yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D37589