[BACKPORT 2.20][#22533] YSQL: fix setrefs for index scan
Summary:
Original commit: 141703ac3dc80ad12673881c280311b6b1dc00e5 / D37487
The referenced IndexScan's indextlist wasn't fixed, so in the cases where
IndexScan was in a subquery the Vars in that list pointed to incorrect
RTE after RTEs from the subplan's root were moved to the common RTE
list.

Postgres does not use the indextlist, because all vars in the original
expression lists refer to the main table. It was changed in YB, when
expression pushdown support was added to the IndexScan. The pushdown
expressions associated with the index, should refer to the index relation
instead, and that kind of fix was done. But indextlist was forgotten,
and this list is needed when EXPLAIN looks up the main table columns for
Vars in the Storage Index List.

Also code was refactored to avoid fixing references unnecessarily if
index pushdown is not present.
Jira: DB-11462

Test Plan: ./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressMisc#testPgRegressMiscSerial'

Reviewers: tnayak

Reviewed By: tnayak

Subscribers: yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D37602