[#23543] docdb: Get colocated schema version correctly in WriteQuery
Summary:
For each write op we apply, we check that the schema version of the table is less than or equal to the schema of the write op, but the code to get the schema version of a table was not taking into account whether the table was colocated. For colocated tables, we were just checking against the schema version of the parent table which was always 0 until DB cloning which increased it to 1 during ImportSnapshot (which caused writes with schema_version 0 to fail). This diff changes the WriteQuery code to get the appropriate schema version if the table is colocated.
Jira: DB-12461

Test Plan:
`./yb_build.sh release --cxx-test integration-tests_minicluster-snapshot-test --gtest_filter *CreateTableAfterClone*`
Jenkins: urgent

Reviewers: yyan, mhaddad

Reviewed By: yyan, mhaddad

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D37634