[#22135] YSQL: Avoid read restart errors with ANALYZE
Summary:
In the current state of the database, ANALYZE can run for a long time on large tables. This long duration increases the chances of errors. We want to minimize such error situations since running analyze again when there is an error can be expensive.

First, we tackle the read restart errors. ANALYZE does not require strict read-after-commit-visibility guarantee, i.e. slightly stale reads are not an issue for the ANALYZE operation. Therefore, we want to avoid these errors for ANALYZE in particular. For this reason, we do not use an ambiguity window (i.e. collapse the ambiguity window to a single point) for ANALYZE.

Moreover, in the current state of the database, DDLs are executed in a "special" transaction separate from the usual transaction code path. This means that multi-table ANALYZE operations such as `ANALYZE;` use a single read point for the entirety of the operation. This is undesirable since there may be a lot of tables in the database and that increases the risk of a snapshot too old error. For this reason, we explicitly pass a fresh read time for ANALYZE of each table from the Pg layer to the tserver proxy. Pg does not exhibit this problem since (a) it runs ANALYZE of each table in a separate transaction (b) it does not cleanup MVCC records that are in use.
Jira: DB-11062

Test Plan:
Jenkins

#### Test 1

```
./yb_build.sh --cxx-test pg_analyze_read_time-test --gtest_filter PgAnalyzeReadTimeTest.InsertRowsConcurrentlyWithAnalyze
```

Insert rows concurrently with analyze to trigger read restart errors.

#### Test 2

```
./yb_build.sh --cxx-test pg_analyze_read_time-test --gtest_filter PgAnalyzeReadTimeTest.AnalyzeMultipleTables
```

Analyze two tables and do a full compaction between the two analyze.

#### Test 3

```lang=sh
$ ./bin/ysqlsh
yugabyte=# create table keys(k int);
CREATE TABLE
... concurrently insert rows using ysql_bench and wait for a while
yugabyte=# analyze keys;
... fails with a read restart error prior to this change but not with this change.
```

To insert rows concurrently use the following sql script

```name=insert.sql,lang=sql
\set random_id random(1, 1000000)
INSERT INTO keys (k) VALUES (:random_id);
```

Ran ysql_bench using

```lang=sh
build/latest/postgres/bin/ysql_bench -t 100000 -f ../insert.sql -n -R 200
```

Reviewers: pjain, bkolagani, yguan

Reviewed By: bkolagani, yguan

Subscribers: ybase, smishra, svc_phabricator, steve.varnau, yql

Differential Revision: https://phorge.dev.yugabyte.com/D37648