[#23251] YCQL, ASH: Fix CQL test failures with ASH
Summary:
There seems to be some race conditions in CQL tests due
to D3169, it is not clear from the summary what that diff
was supposed to fix, so this diff reverts the code to
reading and writing into `request_` non-atomically.

Running a mini cluster with both YSQL and YCQL seems
to be slow in TSAN mode. CQL driver seems to timeout
most of the times while connecting to the cluster. This
diff fixes this by using either YSQL or YCQL while running
the tests.

This diff also enables concurrent updates in some cql
wait states, which fixes a few non-ASH tests when ASH
would be enabled by default
Jira: DB-12182

Test Plan: ./yb_build.sh --cxx-test wait_states-itest

Reviewers: amitanand, jason, hbhanawat

Reviewed By: amitanand

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D37312