[#23739] docdb: Add field transaction_aware back to master_backup.proto
Summary:
Commit b1a90b9 removed some fields from `master_backup.proto`, but external clients (e.g.
YBC) that are interacting with an older DB version with the new proto file still need a way to set
`transaction_aware` to true in CreateSnapshot.

**Upgrade/Rollback safety:**
Once YBC pulls this change in, it will be able to set the transaction_aware field again, regardless of what the DB version is. Old versions of the DB (before the field name was removed) will see this field (as intended) because the field number is the same, and ignore it (also as intended) because they do not use it.

Jira: DB-12644

Test Plan: Jenkins: build only

Reviewers: hsunder

Reviewed By: hsunder

Subscribers: svc_phabricator, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D37686