[#23756] Add sleep after pg_yb_catalog_version bump up when connection manager is enabled
Summary:
When ysql connection manager is enabled , it is possible that a single logical client is connected to different backends to perform transactions. Some of the DDLs could bump up the catalog version which require change in system catalog tables. For that change to be propagated from master to tserver via heartbeat RPC response and then available to backends via shared memory there will be delay of 2 * `heartbeat_interval_ms`.

Hence if two DDLs are performed by logical client on different backend within this time interval ,the shared memory will not be updated leading to non-invalidation of stale cache and eventually the DDL operation failing with catalog mismatch error.

This diff adds a sleep of 2 * `heartbeat_interval_ms` if connection manager is enabled and catalog version is bumped up for a database.

Test Plan:
Jenkins: enable connection manager, all tests

  ./yb_build.sh release --java-test org.yb.ysqlconnmgr.TestMisc#testCreateDropIndex

Reviewers: asrinivasan, skumar, rbarigidad, vpatibandla, mkumar, stiwary, myang

Reviewed By: skumar

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D37627