[pg15-upgrade] fix: remove dependency on --include-yb-metadata flag
Summary:
The pg15 upgrade was using the --include-yb-metadata flag for
ysql_dump to enable various YugabyteDB-specific features in the dump
phase of the catalog metadata upgrade. The --include-yb-metadata flag
is intended for backups only and isn't appropriate for the
upgrade. Switch all currently working upgrade functionality to use the
existing --binary-upgrade flag instead, or the new IsYugabyteEnabled
global added in this commit.

Test Plan:
Jenkins

On MacOS 14 arm64:
  ./yb_build.sh release --sj
  pg15_tests/get_shell_test_specs.sh | grep upgrade | pg15_tests/run_tests.sh
  ./yb_build.sh fastdebug --sj
  pg15_tests/get_shell_test_specs.sh | grep upgrade | pg15_tests/run_tests.sh

Reviewers: jason

Reviewed By: jason

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D37710