[PLAT-14959] Disable connect only for onprem
Summary:
Prior to this diff, the connect to node/pod action gets disabled if there is no access key for the universe.
For Kubernetes universes, we don't need access key information for the command to connect to a pod.
Thus, it doesn't make sense to disable the connect to pod action for Kubernetes universes.

This diff adds a check for whether the universe was created with an on-prem provider which specified manual provisioning (`isOnPremManuallyProvisioned`).
This way, only universes using on-prem providers with manual provisioning would have their connect to node action disabled if no access keys are present.
No access keys could be present in this scenario as access keys are optional for on-prem providers with manual provisioning.

Test Plan:
Verify the connect option is not disabled for k8s pods.
{F283154}

Reviewers: vbansal, rmadhavan, kkannan, lsangappa

Reviewed By: rmadhavan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D37874