[PLAT-15212] Failover local provider test should validate safetime is restored during failover
Summary: Modify DR failover local provider test to make sure safetime is restored as expected. We simulate a row insert after the safetime we get. When we restore to this safetime, it will not contain this row as expected. Need to wait for https://phorge.dev.yugabyte.com/D37310 for fix

Test Plan: Local provider test succeeds.

Reviewers: hzare

Reviewed By: hzare

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D37921