[PLAT-15228] Update package installation command for YBA
Summary:
We are seeing a dependency issue with react-bootstrap-table using an older version of react in its peer dependencies.
Given we don't have any plans to upgrade react-bootstrap-table soon and the fact that we want to keep our react version on 17, it was decided that we will use the `--legacy-peer-deps` option to install the packages anyways.

Test Plan:
Tested on a VM that replicated the dependency resolution issue seen in the build environment.
Observed a succesful `yb_release` run with no npm error.

Reviewers: rmadhavan, kkannan, lsangappa

Reviewed By: rmadhavan, lsangappa

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D37923