[PLAT-14888]: Fix old dr config which does not contains pitr params
Summary: In Old Dr configs, pitr params is null which is not allowed with primitive dataType long, so switched to Long object.

Test Plan:
Tested manually that we can load old dr configs which does not have pitr params store and perform set new database, switchover.
Also, verifed that we can edit to set them and they are used if we set another database or perform switchover.

Reviewers: jmak, hzare, cwang, sanketh, #yba-api-review!

Reviewed By: hzare

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D37928