[#23547] YSQL: fix pg_hint_plan crash with pg_hint_plan.enable_hint_table enabled
Summary:
When hint_table is enabled, pg_hint_plan looks up the hint_table using the the current query string as the key, which is done in the extension callbacks: `post_parse_analyze_hook` and `planner_hook`.

The query string is available in the `ParseState` when it is in `post_parse_analyze_hook` as it is passed down as an argument.

On the other hand, `planner_hook` in **pg11** does not have access to the current query string, and `get_current_hint_string` was assuming that it's processing the top-level query.  However, that is not always the case, and it would crash while normalizing the query if the source location of some `Const` nodes in the query tree is beyond the end of the query string.

The solution is not to use the top-level query string if `ParseState` is unavailable and `stmt_location` and `stmt_len` in the Query node indicate that it may not be the top-level query but a subquery fragment or a query in a stored procedure.

**pg15 note: the change to `pg_hint_plan.c` is not applicable to the pg15 variant because the planner_hook receives the correct query string along with the parse tree.  No `get_current_hint_string` function there.**
Jira: DB-12465

Test Plan:
  ./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressThirdPartyExtensionsPgHintPlan'

Jenkins: test regex: .*TestPgRegressThirdPartyExtensionsPgHintPlan

Reviewers: tnayak

Reviewed By: tnayak

Subscribers: smishra, fizaa, yql

Differential Revision: https://phorge.dev.yugabyte.com/D38018