[#20335] DocDB: Use MonoClock for write query metric
Summary:
`CoarseMonoClock` time has been found to move backwards in time. This is likely due to its reliance
on clock `jiffies` which can overflow, and by default is set to `INITIAL_JIFFIES` to force an overflow
5minutes after the machine boots up.

The write query metric `kQlWriteLatency` requires microsecond precision which is not provided by `CoarseMonoClock`.
This diff switches the write query duration to use `MonoClock` which provided nanosecond granularity.
This is not a frequent operation call so the impact from this is minimal.
Jira: DB-9320

Test Plan: TestMonoTime.DontMoveBackwards

Reviewers: rthallam, qhu, esheng

Reviewed By: rthallam, esheng

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D37866