[#23864] DocDB: Move cluster_uuid to server common flags
Summary:
Even though this flag is not used by the tserver itself, it is useful for test and call_home tracking.

The flag used to show up in the LTO builds since yb-master and tserver are the same application in this mode. Commit 26f4851449de8d8b07b86f2cc7cab2f901fff17b/D36830 removed flags that are not used by the process from the /varz endpoint. This change adds it back.

Fixes  #23864
Jira: DB-12771

Test Plan: Jenkins

Reviewers: sanketh

Reviewed By: sanketh

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D37946