[PLAT-15079] Treat dropped on target tables as unconfigured but preselected
Summary:
**Context**
We have added a change to set DroppedFromTarget status for tables which are dropped from
the target universe only here: 1655e697d49f57cf4225da78af6158c3c89cddc4
On the YBA UI, we need to be including these tables in our bootstrap requirement check
because they will require a bootstrap to be added back to replication due to missing table
on target.

**Change**
On the YBA UI, we will consider tables which are dropped on the target as unconfigured,
but preselected.
This means we will preselect the table when users go to the edit table modal, but also
treat the table as unconfigured and send it along for the need bootstrap check.
This will allow the backend to return the detailed need bootstrap response for these
tables which will be presented on the bootstrap summary page.

This diff also hides the 'no full copy planned' category box if there are no groups of tables which meet this category definition.

Test Plan:
Create an xCluster config.
- Drop a table only on the target universe.
- Add an empty table to both the source and target universe.
- Add the new empty table to replication.
- Verify that the bootstrap summary page indicates we require a full copy (aka
bootstrap) for the table which was dropped only on the target.
{F284195}
{F284243}

Reviewers: vbansal, hzare, cwang, rmadhavan, kkannan, lsangappa

Reviewed By: hzare

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D37949